Add Termination Handles to a CommandResult
Summary:
This associates cleanup information with a `CommandResult`. This allows any Runner to return any asynchronous teardown that needs to be completed.

For example the video recording command could return a handle that would stop the video encoding. This handle will then be called just before `fbsimctl` terminates. This would make the need for calling `record stop` unnecessary as `fbsimctl` would handle this for you. If `record stop` is not called before the `fbsimctl` process terminates, the video may not be fully available, as the encoder needs to be told to flush it's buffers. The same is true of `launch_xctest`.

Reviewed By: asm89

Differential Revision: D4697580

fbshipit-source-id: 8c84095675b82f1a0cbd812706534dab88c3d720