Conform Device and Video Encoding to FBVideoRecordingSession
Summary:
Now, the starting of a video recording will be terminatable based on calling an object alone. This means that the following:

`fbsimctl record start -- listen -- record stop`

Can be replaced with:

`fbsimctl record start -- listen`

As the teardown of `fbsimctl` will ensure that the encoder is stopped. It was previously required to chain this additional command so that the Video Encoder had time to flush it's (asynchronous) writing to a file. This is very surprising at first as it means the details of video writing are leaking out into the `fbsimctl` API.

Reviewed By: marekcirkos

Differential Revision: D4697594

fbshipit-source-id: 30a814996a621c81c7c4d4fe1e3c9ec6f6da2eb2