Reduce OS Version Variants to one Class
Summary: These are currently all part of a cluster to ensure that we have one way of declaring these kinds of properties. However, this is very fragile and can break during new Xcode bumps. We can instead use an enum to contain the version strings, and derive others from this.

Reviewed By: marekcirkos

Differential Revision: D4717514

fbshipit-source-id: 481052298c8106d9f2c00513e624b1303c985df1