Consolidate CLI Bootstrapping
Summary: This represents no behavioural change, but instead consolidates the bootstapping so that all entry and exit is the same. This also consolidates the creation of a logger and reporter, to make it far easier to modify later. The main reason for doing this is to change the file handles that reporting occurs on so that `stream` stdout is not mixed with the bitmap data returned

Reviewed By: marekcirkos

Differential Revision: D4745377

fbshipit-source-id: c65c0e231ff34e462d78c6f53b3f56d9658f93ea