Fix build error on Xcode 8.2
Summary:
I previously ran the `CLIBootstrapper` change on Xcode 8.3 betas. It looks like the type inference is better on the `swiftc` on this version of Xcode, as on Xcode 8.2 it doesn't think that the `catch` is exhausive on an existential type. This fixes this by adding an "Unknown Error" wrapper.

Having said that, the "Unknown Error" message should *never* happen. All errors thrown *should* conform to the `Error` protocol and conform to `CustomStringConvertable`

Reviewed By: zsol

Differential Revision: D4754176

fbshipit-source-id: bac3a98822bbe9d806e4b38709b14067ec5e10ee