Add a 'listen' at the end of a sequence of Actions if none provided
Summary: The requirement that a `-- listen` be provided after some asynchronous action is a big shorcoming of the cli. The reason for it existing is that some actions such as `record` or `stream` will perform some work asynchronously and then return control. In order for this asynchonous work to continue, the `listen` needs to be applied otherwise `fbsimctl` will terminate and the asynchronous work will end.

Reviewed By: marekcirkos

Differential Revision: D4778722

fbshipit-source-id: e862739eeed49e0f8614a08a6b94e7c8ffe2a3e5