Use Abbreviated --format= argument
Summary: In hindsight, having lots of possible flags to alter the target format output was a pretty bad idea, with no discernable benefit. For the most part, the default is sane. If you do want to extract certain fields, it may be much easier to just use `--json` instead (`jq` can also be very helpful here).

Reviewed By: zsol

Differential Revision: D4802110

fbshipit-source-id: b24c275bfb15eec858346d0151f03013df9bc818