Add FBTerminationAwaitable
Summary:
Augments `FBTerminationHandle` with additional semantics for checking whether the termination handle has already terminated. The main reason for doing this is it will be possible for consumers (e.g. `fbsimctl`) can respond to this and break out of a `listen` busyloop.

The end-result will be that `fbsimctl launch_xctest` will be usable without an internal timeout. The implicit awaiting on the asynchronous action will finish as soon as the test is finished.

Reviewed By: marekcirkos

Differential Revision: D4809780

fbshipit-source-id: 1dec0d75aa4d1dae754d9b17b477a87be90ca676