Add FBiOSTargetAction
Summary:
Adds a concrete protocol for defining objects that can be used to fully-serialize an Action to be performed against some `FBiOSTarget`.

By doing this we will be able to pass a full blob of JSON to the Frameworks, with this they can perform some action and report success/failure. This has previously been handled only in `fbsimctl`, but I want to make this a feature of the Framework. This means that many of the features of `fbsimctl` will be able to be used in any executable that links the Frameworks.

Reviewed By: antiarchit

Differential Revision: D4786670

fbshipit-source-id: 12c8fc1f2b69516f9e695608bfd08ccc6b747922