Add FBiOSActionReader
Summary:
Adds the `FBiOSActionReader` class.

This is responsible for reading from a socket/file and then passing on the input actions to a target. This will eventually take the place of the `fbsimctl` socket reader implementation and provide more functionality. For now this just uses delegation to provide information about the running of an Action, as well as recieve output to write back

Reviewed By: marekcirkos

Differential Revision: D4850421

fbshipit-source-id: f27930773057e096742c508742e97ff8aafdd813