Support Logging to file on macOS 10.12
Summary:
The migration to `os_log` in macOS 10.12 claims to bridge `asl` to `os_log` based functions. However, there is no equivalent API to `asl_add_output_file`. I imagine this is why logging to a file additionally is broken on macOS 10.12.

This change allows for loggers to be combined by default, so that the correct behaviour is used based on the OS Version. This change does not introduct a new `os_log` logger. Instead, it adds a File Handle based logger this is composed with an `asl` logger to provide output to the logging infrastructure as well as a file.

Reviewed By: marekcirkos

Differential Revision: D4883184

fbshipit-source-id: 7417c5d1b55ba9b02265dc7306f80d931edcd8eb