Predicatable Callback semantics for FBFramebufferSurface
Summary:
The current behaviour here is unpredicable, meaning that synchronization guarantees need to be made by consumers of the API. The current behaviour is that.

1) If there is a surface immediately available when `-[FBFramebufferSurface attachConsumer:]` is called, the consumer is called synchronously
2) When `SimulatorKit` has a surface available it will call it from the calling thread.

This means that the callbacks from 1 and 2 can race. Using a queue and providing the behaviour from #1 as a return value will make the contract far clearer. It also means that consumers of the API may not need additional synchronization.

Reviewed By: marekcirkos

Differential Revision: D4921032

fbshipit-source-id: d77bd0f86a9c068ec7a67ba97a2cd8d5de5a3dc0