Fully Immutable FBXCTestConfiguration
Summary: Previously this class would 'load' itself after it has been initialized. This means that we have to hide this immutability in the interface. Instead of doing this, we can instead do all the parsing up front in a static parse method. This will enable a single-constructor being used for instantiation, when it comes from a JSON deserialization for example.

Reviewed By: marekcirkos

Differential Revision: D4937899

fbshipit-source-id: 0ef723117e0c013643db2de74e49038dc96415e5