Allow passing an FBSimulator from above to FBXCTestRunner
Summary: Increasing the flexibility of the test running that `fbxctest` implements should mean that it is possible to pass a pre-booted Simulator into a test run. This will allow the optimistic booting of Simulators for Application tests. The current runner will always fetch or create a Simulator. We should be able to pass down the fetcher instead.

Reviewed By: marekcirkos

Differential Revision: D4953626

fbshipit-source-id: 3ebe0459b3abefb7e12d2ba3b55ec828f88c377a