Add FBUploadBuffer
Summary: Adds a new kind of buffering system that can be used for the transfer of binary data. Without requiring a state machine, the buffer will consume input data and use this to write to a file location. Additionally, any "excess" data will be chopped off so that it can be re-interpreted by the caller. This will enable the transfer of a binary over a socket. In future gzipping can be added to allow for the transfer of an entire .app file for example. Clients can use the file path to then reference a path on the remote end, that can be used. For example, this can be used to build the transfer of a `.app` over a socket and following on from this, the launch of that `.app`.

Reviewed By: zsol

Differential Revision: D5145784

fbshipit-source-id: 2e4d4bd8dd8cc6a3a18a59cb8f206c1a0d3b16ae