Add Trimmed-Down otest-shim to fbxctest
Summary: Takes `otest-shim` from `xctool` and moves the core behaviour into `Shimulator`. Also adds a new `Maculator` target, which can be used for shimming `XCTest` in the same way as iOS Simulators. Uses environment variables to switch behaviours instead of having 4 shims as `xctool` did. There's also a bunch of pairing down of shims that can be done as `OTest` is something that we no longer need to care about.

Reviewed By: antiarchit

Differential Revision: D4818532

fbshipit-source-id: 6098a5dd00e96fabb5b89895e43e9785012da679