Default 'persist' to NO on Xcode 9
Summary:
This is a change to make the default behaviour of Simulators that are launched headlessly to not be persistent. This will be made optional in #356. By making this change we keep the boot behaviour consistent over Xcode versions.

One reason that this change is important is that killing the parent process that is holding onto a "death trigger port" will cause the `launchd_sim` to teardown. These semantics are useful for `fbxctest` which uses this to prevent Simulators becoming "abandoned" on a host and taking up resources that have to be manually managed.

Reviewed By: antiarchit

Differential Revision: D5273798

fbshipit-source-id: 07d8572bb757667fb3eb7ef8017e98d2903a42dc