Limit Exposure to SimulatorKit in FBSimulatorControl
Summary: `SimulatorKit` is bundled with Xcode and also links Swift runtime overlays. In order to limit this exposure across the frameworks, I want to remove the abandoned references to `SimulatorKit` that are no longer needed as they are handled in the `Framebuffer` directory. It should also be possible to extract this to a sub-framework and have a `CoreSimulator`-only framework and one that loads `SimulatorKit`

Reviewed By: antiarchit

Differential Revision: D5283050

fbshipit-source-id: 0abe16ab63492cf26b646c5685980912a2077dad