Fix SimDeviceIOClient Usage
Summary:
This method signature has changed in Xcode 9. Fortunately, we can test for this before calling it. As a result, there a minor switch that we have to do.

What else has changed is the fact that an `IOSurfaceRef` is returned instead of an `xpc_object_t`. As a result, we have to inspect the object before returning it to the caller. This is also now shared in the callback and the immediate fetch. Added a few comments about what the expectations of retain counts are.

Reviewed By: zsol

Differential Revision: D5309786

fbshipit-source-id: 847fd7fcb5c1da2f0f42ed5eff5ddad8fb390e71