Fix Bitmap Stream surface consumption when there is only one surface
Summary:
Currently the behaviour of the `FBSimulatorBitmapStream` means that if the `IOSurfaceRef` is returned immediately and the callback `didChangeIOSurface:` is never called, then there will be no streaming.

This is the case in Xcode 9, which will call the damage rect callback, but not require the usage of a new `IOSurfaceRef`. We have a queue for all of the writes so can schedule the work there.

Reviewed By: zsol

Differential Revision: D5310066

fbshipit-source-id: 68ed0a416047dc5d1aae9a67a8b49572a97be623