Use SimulatorKit.SimDeviceLegacyClient
Summary: Adds usage of the "Legacy Client" which `SimulatorKit` exposes. This has the advantage that a `HID` can be connected to an already booted Simulator. Since the creation of `IndigoMessage`s and the sending of them are orthogonalized, it is possible to mix and match them. Only sharp edges are that the buffer must be owned for the client as it is delivered asynchronously and we don't want the deallocation of the data to race with the delivery of the message

Reviewed By: zsol

Differential Revision: D5328454

fbshipit-source-id: 73b55f1ac70879d43e1b93e31f8e473c66c5b1dd