Don't force-close file handles on dealloc
Summary:
This can result in `stdout` or `stderr` getting inadvertantly closed.

If the user of this API wishes for the FileHandle to be closed when the FileHandle is no longer referenced, the `closeOnDealloc:` method can be used on `NSFileHandle`, which provides reference-semantics for the termination of a file handle. Manually calling `consumeEndOfFile` also has the same effect.

Reviewed By: c-ryan747

Differential Revision: D5443763

fbshipit-source-id: 986eb401d6ba0e9215235472e27b1d77f7b90cdc