Use nameForLegacyDescription in test listing
Summary: When targeting a specific test case in a Swift test class, the name of the module must be included. However the output of test listing does not include this currently. Therefore we should include this description instead so that the output of test listing can be used to target individual tests when filtering.

Reviewed By: antiarchit

Differential Revision: D5516178

fbshipit-source-id: e6068f75cc8ce9db83e1df1ba8305d4e1eee6957