FBTask Output is a mirror of the Input
Summary:
Currently all of the `FBTask` processes will accumilate output in some way in order that a string representation can be made. This may be convenient for some cases, but does mean that data is accumilated, either in-memory or on disk. For long running and very chatty processes, this is a huge resource burden.

Instead, the output of the `FBTask` will be a mirror of the requested input types in `FBTaskConfiguration`. This also becomes available in the relevant slots of the `FBTask` implementation. This also makes it easier to specify when data is preferred as output in preference to a string.

Reviewed By: c-ryan747

Differential Revision: D5562207

fbshipit-source-id: cf28202a630fdffcedaedd60ae8076ab8fcea6e4