Don't automatically require Framebuffer connection on --direct-launch
Summary: This is no longer a requirement as of Xcode 8. This means we can do less work on a Direct Launch and leave any overriding of connecting a Framebuffer to `FBSimulatorBootStrategy` which will make any necessary changes. Default boot configurations can be provided inside `FBSimulatorControl` itself if a framebuffer is required to be connected and there is no Framebuffer configuration.

Reviewed By: c-ryan747

Differential Revision: D5591789

fbshipit-source-id: 3df1f91959061c878b09e00449c938b8de47ab20