Remove +[FBXCTestContext contextWithSimulator:]
Summary: This was originally made in order to inject a Simulator into a runner, when not used from `fbxctest`. This is not needed, as each of the runner classes can themselves take Simulators in the constructor. In some cases, such as for logic tests, the Simulator is completely abstracted via an Executor.

Reviewed By: antiarchit

Differential Revision: D5697132

fbshipit-source-id: 3af52860763972a15e75a3747254d1e295561a59