Use NSFileDescriptor instead of file handles in FBControlCoreLogger
Summary: One of the advantages to using an `NSFileHandle` instead of a plain integer file descriptor is that a `NSFileHandle` has reference semantics, which means when the reference count for it hits zero it will deallocate. When injecting a File Handle into a Logger, this means that the `NSFileHandle`'s lifecycle needs to be managed externally. Instead, the `NSFileHandle` can be passed to the logger with the knowledge that it will be closed when the logger is no longer referenced.

Reviewed By: c-ryan747

Differential Revision: D5697254

fbshipit-source-id: f3615c729044d5075080d90aa07771ce3a8d18b9