Add FBFuture
Summary: Futures are surprisingly easy to implement and will make the API a lot more friendly for many cases. We're already half of the way there with `FBTerminationAwaitable` so we may as well go all-in as this will allow for a much nicer API to use and combine with other futures.

Reviewed By: c-ryan747

Differential Revision: D5619700

fbshipit-source-id: 8c4b3072256a8ce2991fcb0a79f69940c4ce4d72