Add a queue for async operations
Summary: Adds another queue that can be used for concurrent operations on an iOS Target. This can be used for operations that don't depend on mutating, or viewing the state of the iOS Target. For example, this can be used for performing transformations on data or reading a file that is only used in one operation.

Reviewed By: c-ryan747

Differential Revision: D5745198

fbshipit-source-id: 6cae6486d25b04479a0083748d53a8d83ec44b91