-[FBBatchLogSearch searchOnTarget:]
Summary: This will allow for future optimizations and changes including making the log searching asynchronous, so that it does not block the calling thread. As a nice side-effect the `fbsimctl` implementation becomes smaller and can also work across Simulators/Devices. There's also now a notion of a timeout so we can prune expensive operations that would otherwise block all other activities

Reviewed By: perryjrandall

Differential Revision: D5745164

fbshipit-source-id: bb9de96e8fe5b860c58d45053d397e7243f03a81