Bridge FBSimulatorAgentOperation to FBFuture
Summary: This adds the ability to resolve an `FBSimulatorAgentOperation` as a future, which then becomes easier to chain and await the completion of. By doing this, we will be able to spawn asynchronously, with the chaining of actions on completion as well as proper failing semantics if the process fails to execute.

Reviewed By: perryjrandall

Differential Revision: D5745165

fbshipit-source-id: 1691670969a6a59c3a303755294d416ad7c52807