Don't reinstall apps on launch
Summary:
At the moment for UI tests you have to provide a path to the app to test thats not installed on the target. This is because every time it attempts to launch a UI test it reinstalls the app.

So I've added in a check to avoid reinstalling the app if the path given is to the installed one

This has the benifit of saving the app install time for each test run in a UI test bundle (in a small example this halved the total execution time)

Reviewed By: lawrencelomax

Differential Revision: D5756368

fbshipit-source-id: edf82befbe48660df96ba22aaba93cb300e0e526