Return futures in FBTestBundleConnection
Summary: Now that we have futures, we can return them and access them in a thread-safe manner whilst the work that occurs happens in a background queue. This introduced them to `FBTestBundleConnection` so that we can flatten out a lot of the work with the removal of calling back to a target's serialization queue.

Reviewed By: c-ryan747

Differential Revision: D5619829

fbshipit-source-id: 6b3bd32beb215f84ba230f91b975861bf08614e3