Fix FBSimulatorTestRunStrategy Contract
Summary:
The method is currently called `connectAndStartWithError`. This is not currently the case as there is a separate `execute` future that corresponds to the `connect` future. This is correctly called in `FBApplicationTestRunStrategy` but not in other code paths. Longer-term these paths need to be consolidated, however this change will at least fix the `FBSimulatorXCTestCommands` case.

Additionally, the futures for execution are now memoized. The reason being that if there are multiple callers of `execute`, they will share the same resolution. Only one execution is currently supported, so this is exepected. This prevents `execution` being called twice, if there are multiple callers of `execute`

Reviewed By: c-ryan747

Differential Revision: D5823151

fbshipit-source-id: 48368cc61621663e5eb22e69c7d1132d79a06807