Add -[FBFuture race:]
Summary: This allows, amongst other things, the ability to have a timeout future that will cancel all other pending operations. This is paricularly useful as it means that timed-out futures do not have to have an awareness of firing a timer of thier own, this is then composed with a `race:` future.

Reviewed By: c-ryan747

Differential Revision: D5786574

fbshipit-source-id: 1cb6dfaa1bd84ba9410986e03bbd7f83dc0eb49e