Make FBXCTestProcess Futures-Based
Summary: We have a bunch of waiting and polling ocurring inside the execution of an `xctest`. This makes it very difficult to use and impractical outside of the main queue. By moving to futures, we can control on which queues the work is progressing and make the process of error handling substantially easier. There's a lot of code removal here, due to the simplification of the interface and all the platform-specialization provided in the contract of a Future.

Reviewed By: c-ryan747

Differential Revision: D5793412

fbshipit-source-id: 1b4aee4acdbdc724e242526e5fd3c37c064bbbe7