Remove Unused FBSubject Classes
Summary:
Anything in `FBControlCore` land that is reported will conform to `FBJSONSerializable` and be an `NSObject` so therefore conforms to `CustomStringConvertible`. This means that all of these classes are nominally `ControlCoreSubject` types in Swift-land.

In the case of the `RecordSubject` this will be represented with a `FBiOSTargetAction` at some point, rather than having a specialized `Subject` class. This is the first step in hiding Subject classes that *cannot* be reported to an `FBJSONEventInterpreter` without hitting an assert as `EventName` or `EventType` is missing.

Reviewed By: antiarchit

Differential Revision: D5451354

fbshipit-source-id: a0b451f9a5c0de6388a081e1801e1e799dd1b5ce