Make -[FBFuture await:] have an forever timeout
Summary:
In the effort to bridge a large number of synchronous APIs towards asynchronous variants, there are a number of levels at which refactoring yeilds awaiting of a future.

When a leaf node in a refactor becomes an `FBFuture`, it may still be wrapped in and composed-with synchronous APIs. The synchronous calls don't have implicit timeouts so will only return when the action has been resolved. When these synchronous calls are replaced with asynchronous ones, they are wrapped with an `await` to extract the value. This is a temporary proposition whilst other APIs are migrated. Having a timeout for the await that is less than "forever" will mean that synchronous API may start timing out unexpectedly, which is very surprising.

In order to counteract this, setting an indefinate timeout on the default `await:` means callers can opt into more agressive timeouts. Eventually, timeouting can be placed on the composition of an API at the root future, for example in the `fbsimctl` cli, or some UI.

Reviewed By: antiarchit

Differential Revision: D6021484

fbshipit-source-id: 3343c37b3b032c72994826ac21b83251561a4e3b