Replace EventReporterSubject definition with FBControlCore one
Summary: The final set of classes that have an `FBControlCore` replacement that is not in `fbsimctl` currenty. This does require some bridging of types that do not have a representaiton in `FBControlCore` yet, but this can be removed over time. This has removed nearly all specialization of event reporting from `fbsimctl` and into the framework.

Reviewed By: antiarchit

Differential Revision: D5985051

fbshipit-source-id: 554ee65d2fbe8552f4062d580a956ab47f989427