Add FBiOSTargetFuture
Summary:
I eventually want to move all `FBiOSTargetAction` instances to use this. This diff adds a way of defining a behaviour similarly to `FBiOSTargetAction` except that a future is returned.

This makes the process of reporting cohesive, as it sits alongside the invocation of the underlying future. Another nice effect of this is that reporting can be decorated on top of the Future, the definition of an action is not far of the API used to get the base future, then further decorated with reporting.

However, it's not a simple task to convert all `FBiOSTargetAction` instances *at once*. It makes more sense to migrate existing `FBiOSTargetAction` classes slowly. Any new Actions that come across can start-off as `FBiOSTargetFuture`s and never need to use the old sychronous way of doing this.

This does mean that a "bridge" is required at some level in order for a `FBiOSTargetFuture` to be defined in the synchronous way. This is also another crutch that can be remove. Any future running integrators can start using futures from the beginning.

Reviewed By: antiarchit

Differential Revision: D5910509

fbshipit-source-id: 123346af5bd0ce67a8e2afea6d94251273639f2e