Add Future Resolution on HttpRelay
Summary: We absolutely do not want to perform a synchronous dispatch to the main queue, as this will prevent any further servicing of the main dispatch queue, including via run-loop spinning. Moving to a model where the synchronous waits are performed at the root level will mean that more and more work can be made truly asynchronous and there is only wait on the result that we care about.

Reviewed By: antiarchit

Differential Revision: D5880556

fbshipit-source-id: a8e25a6e6c0660ea1d71bc520ff9c92352f34f3b