Resolve work from HttpRelay on non-main queue
Summary:
Hidden inside the quirks of the implementation of current implementation of some synchronous APIs is the following behaviour:

1) A synchronous call makes a call to an async API from the main queue.
2) In order to get some async data, or confirm that the async API call has finished, the run loop is spun
3) Spinning the Run Loop does more than just wait for an async condition to be fulfilled, it cranks the Run Loop's input sources so that any `dispatch_async`'d calls to the main queue. Each dispatch queue can have it's own run loop but the main queue is specialized. You can see this when `__CFRUNLOOP_IS_SERVICING_THE_MAIN_DISPATCH_QUEUE__`` is in the backtrace.
4) This causes the `dispatch_async`'d calls to the main queue to be resolved. In some cases this servicing of the main dispatch queue can result in the fulfillment of #2

This seems fine, but means that there are implicit dependencies on the main queue being servicable in order to some commands to be fulfilled. One area that this is particularly painful is in how work from the `HttpRelay` is scheduled via the `ActionPerformer` protocol. Currently, the implementation of this relies on data being sychronously returned via the main queue. The main queue is here used as a primitive for synchronization, rather than as a hard-requirement of the implementation of some commands.

Since the main queue is synchronously dispatched to in `ActionPerformer` the duration of a long-running synchronous operation will mean that the main dispatch queue cannot be serviced and so the condition in #2 cannot be fulfilled.

This will also be improved upon with a move to futures, since work can then be awaited from any queue. Additionally, assumptions about the usage of the main queue can also be pruned over time.

Reviewed By: antiarchit

Differential Revision: D6008067

fbshipit-source-id: d2f8a8f0ab6f9bb78598b71d594e59e3fbbeb13f