Future-based File Reader finishing.
Summary:
File Readers terminate asynchronously as queues are used for the serialization of IO operations.

The current implementation does not take the signalling here into account, which could lead to racing if the IO channel hasn't fully drained.

Wrapping this in a future makes it easier to wait for the IO to be fully drained before moving on. This is optional, so that the same behaviour as before is preserved. If you care about waiting, await the future, otherwise just use the same API as before.

Reviewed By: marekcirkos

Differential Revision: D6098582

fbshipit-source-id: 54f791fafe2b6445ffec20db54439b591f605741