Consolidate Approvals in Public Interface
Summary: Now that all these commands are futures-based, we can hide how these are implemented by consolidating them behind the enum. This enables the `approve` action to take switches and therefore operate on each of these in parallel. There are no new-switches in the fbsimctl `approve` command yet, but that can be added by using the enum in future.

Reviewed By: antiarchit

Differential Revision: D5852406

fbshipit-source-id: 3131c68a02f8507774363554d5934b3dd962991b