FBTerminationAwaitable is Futures-Based
Summary:
This is far superior to polling on a `BOOL` and allows reaction to some long-running task being resolved. It also allows `FBiOSTargetFuture` to return something different than just a `FBiOSTargetActionType`. This will then enable actions to define their own ways of building a long-running task that can be awaited.

At the moment this can be done by passing a termination awaitable into the delegate, but in future we will be able to return a long-running task, directly from the invocation of an `FBiOSTargetFuture`

Reviewed By: marekcirkos

Differential Revision: D6171506

fbshipit-source-id: 958dc30418d5c1f36103d2283b1c2bc95b01ccc3