Make FBXCTestCommand Execution Futures-Based
Summary: Test Operations are now both futures in the starting of a test execution and in the waiting for a test execution to end. This means that the launch is fully asynchronous and the finishing of the test operation can now be awaited in a separate place to where the test is launched.

Reviewed By: marekcirkos

Differential Revision: D6196208

fbshipit-source-id: cc0199348b0bd8bcd0b014a6341054691021cc07