Make FBTestReporterForwarder strongly-reference reporter
Summary: In `FBSimulatorTestRunStrategy` falls out scope and deallocates it's `reporter` argument, then this can be nullified in the `reporter` reference of `FBTestReporterForwarder`. The result of this is that the reporter may not recieve all messages. The Forwarder should strongly reference here, `weak`ing doesn't prevent a retain cycle, because no-retain cycle exists to be broken.

Reviewed By: marekcirkos

Differential Revision: D6263336

fbshipit-source-id: 3d4fe8556bf0b4e6d4d967a9c09006d8361e7a06