Make FBSimulatorBootVerificationStrategy Aynchronous
Summary: This strategy has a run-loop spin in order to verify the availability of services. The underlying APIs are all futures based, and we can compose a future that runs in a loop until some condition is met. This will have the effect that the calling queue is massively freed whilst the IO occurs.

Reviewed By: marekcirkos

Differential Revision: D6123206

fbshipit-source-id: 1e4dd6f8f4484cfe93140127c2137e7caf449854