Replace FBiOSActionAwaitableDelegate with FBiOSTargetContinuation
Summary:
Always be returning™. Instead of returning the handle alone, we can change a `FBTerminationAwaitable` being returned in the future, with an optional future to wait for. This removes the need for wrapping with delegation and simplifies callsites.

If the caller is interested in waiting for the liecycle to finish, it can do. Otherwise it can ignore this and continue. This also means that the `FBTerminationAwaitable` can become specialized as the return-value of an `FBiOSTargetFuture`. Eventually we can remove the dependency on `FBTerminationHandle`.

Another effect is that we no longer need a component that intercepts delegate methods that report the continuation. This makes the process of delegation substantially easier to work with.

Reviewed By: marekcirkos

Differential Revision: D6311883

fbshipit-source-id: 28259de75bcd20fa2fa0aedb0da693c9be743a81