Remove FBiOSTargetFutureContinuation's dependency on FBTerminationHandle
Summary:
This removes any of the synchronous `-(void)terminate` commands so that any asynchronous termination is truly asynchronous, pushing these semantics into the cancellation of the `completed` future.

One nice property that we get here is the ability to use `FBFuture.race` to cause cancellation of the Signal-Handler vs. the Continuation. Whichever occurs first will cancel the other.

Reviewed By: dtorres

Differential Revision: D6311911

fbshipit-source-id: 425181e1c193e0bfb520f24b171f51aa9e4d63aa