Ensure that consumeEndOfFile is called before -[FBFileReader stopReading] has resolved
Summary: Due to the way that `dispatch_io` notifying of us jumbled in with the public API of `FBFileReader`, the ordering of `consumeEndOfFile` would be indeterminate. By having a private Future that resolves when the `dispatch_io` is done and a public Future that is chained on top of this, we can ensure that `consumeEndOfFile` is called *before* `stopReading` resolves.

Reviewed By: dtorres

Differential Revision: D6397093

fbshipit-source-id: 122267c90cffd75a075f7f0d9f564f36cc6cbbe1