Allow EOF consumption to be inspected on an FBLineFileConsumer/FBAccumilatingFileConsumer
Summary: There's currently no way for a consumer block on `FBLineFileConsumer` to know if it has reached an EOF as this is kept private to instances of the class. It can be important to know when this event has occured so that callers can block until all async work has been drained. To expose this without requiring additional blocks or consumers, a `FBFuture` is exposed on the public API. This also makes it easy to combine with existing chains of asynchronous work

Reviewed By: marekcirkos

Differential Revision: D6398179

fbshipit-source-id: 1ba0ca5234e434eb5ed43d00dded4950617962cc