Fix segfault on -[SimVideoFile completionHandler] with dispatch_io
Summary:
It seems that we can't get around the fact that `-[SimVideoFile completionHandler]` can sometimes be a dangling pointer. Not 100% sure why this is the case, but it looks like this is never an issue in `simctl` as the process exits much earlier.

The workaround here is to use `dispatch_io` to be notified of when the file has finished writing and a global block to take the place of the dangling pointer. We still get a callback at the appropriate time when the file has finished writing and since the `completionHandler` is a global, it won't ever be a dangling-pointer.

Reviewed By: antiarchit

Differential Revision: D6404125

fbshipit-source-id: 0255c3393eec0c2aa4b757ff7cae1a1ffff98799