Async Cancellation of Futures
Summary:
Currently a call to a `-[FBFuture cancel]` is synchronous. This means that either 1) The cancellation is blocking on some work, or more likely 2) Any asynchronous teardown is not inspectable. This means that it becomes difficult to ensure that there is async teardown for async setup. Now cancellation can (optionally) return some Future that represents the state of the teardown.

This is great for Video Recording for example. Calling `cancel` will cause the stopping of video recording. If the caller cares about waiting for stopping to complete (in the case of video, it really should), then this too can be inspected.

Reviewed By: marekcirkos

Differential Revision: D6468407

fbshipit-source-id: 80156ae93ce114e198b4b80f0a5e2295c254ad8b