Container App Lifecycle is now O(1)
Summary:
Extracts out the Container App detection to a separate component. This is helpful for a few reasons:
1) Slimming down the event sink composition, there's some bulk in the `EventRelay` that I want to prune
2) Making the whole process O(1) instead of O(n) on the number of Simulators.

Currently, there is a Notification registration *per simulator*. This is equivalent to enumerating the an array of Simulators to obtain the correct one. Instead of doing this, we can make the lookup dictionary based and make the process substantially easier at the same time.

Reviewed By: marekcirkos

Differential Revision: D6475707

fbshipit-source-id: 3d83034d9144ed801207138087e7d40d4a3c8886