Move Synchronous Task Completion to Builder
Summary:
We already have a consistent way of waiting for the completion of a task, and that is to use the completion future. This can now be removed and replaced with that.

For all existing callsites, the convienience is placed on the Builder, as it does not need to be part of the core implementation of the task.

Reviewed By: dtorres

Differential Revision: D6496039

fbshipit-source-id: 14143a1cf39afdca267aaa20467cec5d7b6cd365