FBTask is only returned in a started state
Summary:
Currently, it is possible for the process identifier to be invalid before the Task has started. We already have the `FBTaskBuilder` for setting up a task, so there's no reason to have multi-stage factorying at the `FBTask` instance level. There are also no callers that care about having a Task in a Not-Yet-Started state. We want to have a higher-level API without having to worry about `NSInvalidArgumentException`s throwing all over the place.

This also makes it possible to move the conveniences *outside* of the core `FBTask` implementation. This keeps the `FBTask` as lean as possible and the builder is responsible for adding all of the convenience.

Reviewed By: marekcirkos

Differential Revision: D6555318

fbshipit-source-id: 426991abf962eb20ff3d9feb79c47cf4357ffeba