Use AMDevice Notifications to update AMDevice Reference
Summary:
It's possible that `AMDevice` references get stale at some point, since the `AMDeviceRef` references that are created can change over time for the same UDID.

Here we can change the way that we create `FBAMDevice` wrappers and use the fact that these are Reference types.

Before, `FBAMDevice` classes were created every time the user called `+[FBAMDevice allDevices]`, leading to new references that wrapped the underlying `AMDeviceRef`. Now we ensure that these references are shared for the same underlying device. As these references are now shared they can be updated in response to notifications. From the perspective of the caller, it doesn't need to care about this happening. The references to the `AMDeviceRef` is updated without it's knowledge and all references for the same `FBAMDevice` are shared.

Reviewed By: dtorres

Differential Revision: D6650862

fbshipit-source-id: 0e2f46dd106734de79bee1373e36cf310a78ea4c