Update -[FBDeviceSet allDevices] from Notifications
Summary: Currently this Array of `FBAMDevice` instances is constructed every time that `-[FBDeviceSet allDevices]` is accessed. This results in a lot of repeated work and can mean that the device list lags behind the source of truth via `FBAMDeviceManager`. We have the necessary semantics that `FBAMDevice` references are now globally shared, so we can avoid doing this for the `FBDevice` instance that wraps it in the User-Facing API.

Reviewed By: dtorres

Differential Revision: D6655245

fbshipit-source-id: 85f9f64606f0813aa52c4c2ac0f1680e04c15ebb